sync

+++ project2/02_discNJN_1D/dataPrepare/cB211.072.64/post2avgsrc.py
@@ -6,9 +6,10 @@ import numpy as np
 
 ens='cB211.072.64'
 
-lat_L={'cB211.072.64':64,'cC211.060.80':80,'cD211.054.96':96}[ens]
+lat_L={'cB211.072.64':64,'cC211.060.80':80,'cD211.054.96':96,'cE211.044.112':112}[ens]
 
-max_mom2={'cB211.072.64':23,'cC211.060.80':26,'cD211.054.96':26}[ens]
+# max_mom2={'cB211.072.64':23,'cC211.060.80':26,'cD211.054.96':26,'cE211.044.112':4}[ens]
+max_mom2={'cB211.072.64':1,'cC211.060.80':1,'cD211.054.96':1,'cE211.044.112':1}[ens]
 range_xyz=range(-int(np.sqrt(max_mom2))-1,int(np.sqrt(max_mom2))+2)
 moms_pc=[[x,y,z] for x in range_xyz for y in range_xyz for z in range_xyz if x**2+y**2+z**2<=max_mom2]
 max_mom2=1
@@ -19,7 +20,11 @@ moms_target=[pf+pc for pf in moms_pf for pc in moms_pc]
 moms_target.sort()
 # moms_target=np.array(moms_target)
 
-tfs={'cB211.072.64':range(2,26+1),'cC211.060.80':range(2,28+1),'cD211.054.96':range(2,32+1)}[ens]
+# tfs={'cB211.072.64':range(2,26+1),'cC211.060.80':range(2,28+1),'cD211.054.96':range(2,32+1),'cE211.044.112':range(2,32+1)}[ens]
+tfs={'cB211.072.64':range(2,22+1),'cC211.060.80':range(2,26+1),'cD211.054.96':range(2,30+1),'cE211.044.112':range(2,32+1)}[ens]
+
+stouts=range(0,40+1)
+# stouts=range(0,4)
 
 def src2ints(src):
     (sx,sy,sz,st)=re.search('sx([0-9]*)sy([0-9]*)sz([0-9]*)st([0-9]*)',src).groups()
@@ -33,12 +38,12 @@ def get_phase(src_int,mom):
 @click.command()
 @click.option('-c','--cfg')
 def run(cfg):
-    inpath=f'/p/project/ngff/li47/code/projectData/02_discNJN_1D/{ens}/data_post/{cfg}/'
-    outpath=f'/p/project/ngff/li47/code/scratch/run/02_discNJN_1D/{ens}/data_avgsrc/{cfg}/'
+    inpath=f'/p/project1/ngff/li47/code/projectData/02_discNJN_1D/{ens}/data_post/{cfg}/'
+    outpath=f'/p/project1/ngff/li47/code/scratch/run/02_discNJN_1D_new/{ens}/data_avgsrc/{cfg}/'
     os.makedirs(outpath,exist_ok=True)
     files=[file for file in os.listdir(inpath) if file.startswith('N.h5')]
     
-    inpath_fullmom=f'/p/project/ngff/li47/code/projectData/02_discNJN_1D/{ens}/data_N_fullmom/{cfg}/'
+    inpath_fullmom=f'/p/project1/ngff/li47/code/projectData/02_discNJN_1D/{ens}/data_N_fullmom/{cfg}/'
     files_fullmom=[file for file in os.listdir(inpath) if file.startswith('N.h5')]
     outfile=f'{outpath}N.h5'
     outfile_flag=outfile+'_flag'
@@ -80,7 +85,7 @@ def run(cfg):
         flas=['N_N']
         # js=['j+','j-','js','jc']
         # js=[j for j in fj['data'].keys() if not j.startswith('j-') and ';' in j]
-        js=['j+;g{m,Dn};tl','js;g{m,Dn};tl','jc;g{m,Dn};tl']
+        js=['j+;g{m,Dn};tl','j-;g{m,Dn};tl','js;g{m,Dn};tl','jc;g{m,Dn};tl']
         for j in js:
             outfile=f'{outpath}discNJN_{j}.h5'
             outfile_flag=outfile+'_flag'
@@ -146,12 +151,84 @@ def run(cfg):
                         t=j.split(';')[1:]; t=';'.join(t)
                         ky=f'inserts;{t}'
                     else:
-                        ky='inserts'
+                        ky='inserts'   
                     fw.create_dataset(ky,data=fj[ky][:])
                     for key in data:
                         fla,tf,j=key
                         fw.create_dataset(f'data/{fla}_{j}_{tf}',data=data[key]/Nsrc)
                         fw.create_dataset(f'data_bw/{fla}_{j}_{tf}',data=data_bw[key]/Nsrc)
+                            
+                os.remove(outfile_flag)
+                
+        flas=[f'{stout}' for stout in stouts]
+        js=['jg;stout']
+        for j in js:
+            outfile=f'{outpath}discNJN_{j}.h5'
+            outfile_flag=outfile+'_flag'
+            if (not os.path.isfile(outfile)) or os.path.isfile(outfile_flag):
+                with open(outfile_flag,'w') as f:
+                    pass
+                
+                data={(fla,tf,j):0 for fla in flas for tf in tfs}
+                data_bw={(fla,tf,j):0 for fla in flas for tf in tfs}
+                Nsrc=0
+                for file in files:
+                    flag_setup=True
+                    with h5py.File(f'{inpath}{file}') as fN:
+                        if flag_setup:
+                            moms=[list(mom) for mom in fN[f'moms'][:]]
+                            momMap_N=[moms.index(mom[:3]) for mom in moms_target]
+
+                            # The loop has exp(-iqx) as the phase, the momentum conservation is p'(sink) + q(transfer) = p(source).
+                            # This is opposite to what is used in many ETMC papers.
+                            moms_j=[list(mom) for mom in fj[f'moms'][:]]
+                            momMap_j=[moms_j.index(mom[-3:]) for mom in moms_target]
+
+                            flag_setup=False
+                                
+                        for src in fN['data'].keys():
+                            src_int=src2ints(src); st=src_int[-1]
+                            tPhase=np.array([get_phase(src_int,mom) for mom in moms_j])
+                            Nsrc+=1
+                            # print(Nsrc,end='                     \r')
+                            
+                            for fla in flas:
+                                datj=fj[f'data/{j}{fla}'][:]
+                                datj=datj*tPhase[None,:,None]
+                                for tf in tfs:
+                                    # (time,mom,dirac/proj,insert)
+                                    tN=fN[f'data/{src}/N_N'][tf,:]
+                                    tN=tN[momMap_N]
+                                    tN=np.transpose(tN[...,None,None],[2,0,1,3])
+                                    # print(tN.shape)
+                                    tj=np.roll(datj,-st,axis=0)[:tf+1]
+                                    tj=np.transpose(tj[:,momMap_j][...,None],[0,1,3,2])
+                                    # print(tj.shape)
+                                    data[(fla,tf,j)] += tN*tj
+                                
+                                    # (time,mom,dirac/proj,insert)
+                                    tN=fN[f'data_bw/{src}/N_N'][-tf,:]
+                                    tN=tN[momMap_N]
+                                    tN=np.transpose(tN[...,None,None],[2,0,1,3])
+                                    # print(tN.shape)
+                                    tj=np.roll(datj,-st-1,axis=0)[::-1][:tf+1]
+                                    tj=np.transpose(tj[:,momMap_j][...,None],[0,1,3,2])
+                                    # print(tj.shape)
+                                    data_bw[(fla,tf,j)] += tN*tj
+                                        
+                            # if Nsrc==10:
+                            #     break
+                            # break
+                
+                with h5py.File(outfile,'w') as fw:
+                    fw.create_dataset('notes',data=['time,mom,proj,insert','mom=[sink,ins]; sink+ins=src','proj=[P0,Px,Py,Pz]'])
+                    fw.create_dataset('moms',data=moms_target)
+                    ky='inserts;jg'
+                    fw.create_dataset(ky,data=fj[ky][:])
+                    for key in data:
+                        fla,tf,j=key
+                        fw.create_dataset(f'data/N_N_{j}{fla}_{tf}',data=data[key]/Nsrc)
+                        fw.create_dataset(f'data_bw/N_N_{j}{fla}_{tf}',data=data_bw[key]/Nsrc)
                             
                 os.remove(outfile_flag)
